feat: implement DELETE method
