+++ src/index.ts
@@ -1,12 +1,12 @@
 import * as http from 'http'
-import * as repository from './repository';
+import * as store from './data';
 import { validate } from 'uuid';
 
 const PORT = process.env.PORT || 5000;
 
 const server = http.createServer(async (req, res) => {
     if (req.url === '/api/users' && req.method === 'GET') {
-        const users = await repository.getUsers();
+        const users = await store.getUsers();
         res.writeHead(200, { 'Content-Type': 'application/json' });
         res.end(JSON.stringify(users).replace(/]|[[]/g, ''));
     } else if (req.url === '/api/users' && req.method === 'POST') {
@@ -18,7 +18,7 @@ const server = http.createServer(async (req, res) => {
       
         const data = Buffer.concat(buffers).toString();
       
-        const newUser = await repository.addUser(JSON.parse(data));
+        const newUser = await store.addUser(JSON.parse(data));
 
         res.writeHead(201, { 'Content-Type': 'application/json' });
         res.end(JSON.stringify(newUser));
@@ -39,7 +39,7 @@ const server = http.createServer(async (req, res) => {
 
         if (req.method === 'GET') {
             try{
-                const user = await repository.getUserById(id);
+                const user = await store.getUser(id);
                 res.writeHead(200, { 'Content-Type': 'application/json' });
                 res.end(JSON.stringify(user));
             } catch (e) {
@@ -59,7 +59,7 @@ const server = http.createServer(async (req, res) => {
             const data = Buffer.concat(buffers).toString();
           
             try {
-                const updatedUser = await repository.updateUser(id, JSON.parse(data));
+                const updatedUser = await store.updateUser(id, JSON.parse(data));
                 res.writeHead(200, { 'Content-Type': 'application/json' });
                 res.end(JSON.stringify(updatedUser));
             } catch (e) {
@@ -70,7 +70,16 @@ const server = http.createServer(async (req, res) => {
             }
         }
         if (req.method === 'DELETE') {
-            
+            try{
+                const user = await store.removeUser(id);
+                res.writeHead(204, { 'Content-Type': 'application/json' });
+                res.end(JSON.stringify(user));
+            } catch (e) {
+                if (e instanceof Error) {
+                    res.writeHead(404, { 'Content-Type': 'application/json' });
+                    res.end(JSON.stringify({ 'error' : e.message})); 
+                }
+            }
         }
     } else {
         res.writeHead(404, { 'Content-Type': 'application/json' });
